replace existing disabled peripheral
